# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "CaseEnvelope": {
        "description": "``Case`` plus the API-shape-only ``_links`` placeholder.",
        "properties": {
          "_links": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Links",
            "type": "object"
          },
          "assigned_to_user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned To User Id"
          },
          "closure_date": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closure Date"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "customer_metadata": {
            "$ref": "#/components/schemas/CustomerMetadata"
          },
          "id": {
            "maxLength": 31,
            "minLength": 31,
            "pattern": "^case_[0-9A-HJKMNP-TV-Z]{26}$",
            "title": "Id",
            "type": "string"
          },
          "latest_decision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CaseLatestDecision"
              },
              {
                "type": "null"
              }
            ]
          },
          "risk_band": {
            "anyOf": [
              {
                "enum": [
                  "low",
                  "medium_low",
                  "medium_high",
                  "high"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Band"
          },
          "state": {
            "$ref": "#/components/schemas/CaseState"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "state",
          "customer_metadata",
          "created_at",
          "updated_at"
        ],
        "title": "CaseEnvelope",
        "type": "object"
      },
...
      "CaseLatestDecision": {
        "description": "Per-case decision summary surfaced on ``GET /v1/cases/{case_id}``.\n\nStory 7.6 / AC #5. ``sealed_ledger_entry_id`` is ``None`` while the\ndecision is still in the 120-second undo window (state\n``pending_seal``); populated once the seal-time ledger entry is\nwritten. Story 7.7 owns the ``decisions`` table that backs this; the\nenvelope addition lands here so the cockpit-ui's\n``SealedIndicator`` can read the seal id without a second round-\ntrip.",
        "properties": {
          "decision_id": {
            "minLength": 1,
            "title": "Decision Id",
            "type": "string"
          },
          "outcome": {
            "enum": [
              "approve",
              "decline",
              "approve_with_conditions",
              "escalate_to_edd"
            ],
            "title": "Outcome",
            "type": "string"
          },
          "sealed_ledger_entry_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sealed Ledger Entry Id"
          }
        },
        "required": [
          "decision_id",
          "outcome"
        ],
        "title": "CaseLatestDecision",
        "type": "object"
      },
      "CaseListResponse": {
        "description": "Pagination envelope per ``architecture.md#Format Patterns``.",
        "properties": {
          "has_more": {
            "default": false,
            "title": "Has More",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CaseEnvelope"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "CaseListResponse",
        "type": "object"
      },
      "CaseState": {
        "description": "Lifecycle states. See ``ALLOWED_TRANSITIONS`` for the edge set.",
        "enum": [
          "intake_scheduled",
          "decision_ready",
          "pending_seal",
          "committed",
          "escalated",
          "closed"
        ],
        "title": "CaseState",
        "type": "string"
      },
...
      "CustomerMetadata": {
        "description": "Customer-side facts attached to a case.\n\n``extra`` is an extensibility escape hatch so fixtures (Story 2-4) can carry\ndemo-only fields without schema churn. Production code should prefer named\nfields over ``extra`` once the bank-buyer scope revives.",
        "properties": {
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "customer_name": {
            "minLength": 1,
            "title": "Customer Name",
            "type": "string"
          },
          "customer_type": {
            "anyOf": [
              {
                "enum": [
                  "individual",
                  "company"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Type"
          },
          "extra": {
            "additionalProperties": true,
            "title": "Extra",
            "type": "object"
          }
        },
        "required": [
          "customer_name"
        ],
        "title": "CustomerMetadata",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Cockpit API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/cases": {
      "get": {
        "description": "Returns every case currently in the cockpit, newest first. Each item carries the case id, current state (intake_scheduled, decision_ready, etc.), customer metadata, and assigned officer. Used by the case_supervisor agent to answer 'what cases are there?'.",
        "operationId": "list_cases_v1_cases_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Cockpit-Demo-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Cockpit-Demo-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List cases (newest first)",
        "tags": [
          "cases"
        ]
      }
    },
...
    "/v1/cases/{case_id}/stream": {
      "get": {
        "description": "Story 4.6 \u2014 yields agent.state_changed / case.state_changed / case.documents_changed events for the given case. Single-worker in-process fan-out (no Redis pub/sub in demo). Auth via ``?as=`` query param (EventSource cannot send custom headers) or the ``X-Cockpit-Demo-User`` header.",
        "operationId": "stream_case_events_v1_cases__case_id__stream_get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "pattern": "^case_[0-9A-HJKMNP-TV-Z]{26}$",
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "as",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As"
            }
          },
          {
            "in": "header",
            "name": "X-Cockpit-Demo-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Cockpit-Demo-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Per-case Server-Sent Events stream",
        "tags": [
          "cases"
        ]
      }
    },
...
  }
}